Optimize service tester script

+++ .github/scripts/service-tester/predefined-service-tester.py
@@ -50,8 +50,11 @@ def test_combination(comb):
     remove(BIN_PATH + "/config.yml")
 
     # Execute Compose Generator with the config file
+    if system(f"cd {BIN_PATH} && docker compose build --quiet") != 0:
+        sys.exit("Docker failed when building the local service(s) for combination " + str(comb))
+
     if system(f"cd {BIN_PATH} && docker compose up -d --quiet-pull") != 0:
-        sys.exit("Docker failed when generating stack for combination " + str(comb))
+        sys.exit("Docker failed when pulling the remote service(s) for combination " + str(comb))
 
     if system(f"cd {BIN_PATH} && docker compose down") != 0:
         sys.exit("Error on 'docker compose down' for " + str(comb))
